AI/requirements.txt
# AI Module dependencies

# Core ML
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Transformers & NLP
transformers==4.37.2
sentencepiece==0.1.99
tokenizers==0.15.1
datasets==2.16.1
accelerate==0.26.1

# Training
peft==0.8.2
trl==0.7.10
bitsandbytes==0.42.0

# Embeddings & Vector Search
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# Web UI for training
gradio==4.16.0

# API clients (for hybrid approach)
openai==1.12.0
anthropic==0.18.0

# Text processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Document parsing
pypdf==4.0.1
python-docx==1.1.0

# Data handling
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
wandb==0.16.3

# Async
aiofiles==23.2.1
httpx==0.26.0
